requirements.txt
djangorestframework
djangorestframework-simplejwt
python-decouple
